chore: reorganize files
